moving it to streamexecution
